+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/AbstractKitTest.groovy
@@ -4,7 +4,6 @@ import org.gradle.testkit.runner.BuildResult
 import org.gradle.testkit.runner.GradleRunner
 import org.junit.Rule
 import org.junit.rules.TemporaryFolder
-import ru.vyarus.gradle.plugin.quality.report.Reporter
 import spock.lang.Specification
 
 /**