+++ src/test/groovy/ru/vyarus/gradle/plugin/quality/tools/pmd/PmdIncrementalAnalysisKitTest.groovy
@@ -10,7 +10,45 @@ import ru.vyarus.gradle.plugin.quality.AbstractKitTest
  */
 class PmdIncrementalAnalysisKitTest extends AbstractKitTest {
 
-    def "Check pmd incremental mode enabling"() {
+    def "Check pmd incremental mode enabled by default"() {
+        setup:
+        build("""
+            plugins {
+                id 'java'
+                id 'ru.vyarus.quality'
+            }
+
+            quality {
+                strict false
+                
+                // does not affect enabled status in gradle  >= 6.4
+                pmdIncremental = false
+            }
+            
+            task validator {
+                doFirst() {
+                    if (!pmdMain.incrementalAnalysis.get()) {
+                        throw new IllegalStateException('PMD incremental analysis not enabled');
+                    }
+                }
+            }
+
+            repositories {
+                mavenCentral() //required for testKit run
+            }
+        """)
+
+        fileFromClasspath('src/main/java/sample/Sample.java', '/ru/vyarus/gradle/plugin/quality/java/sample/Sample.java')
+
+        when: "run check task with java sources"
+        BuildResult result = run('pmdMain', 'validator')
+
+        then: "all plugins detect violations"
+        result.task(":pmdMain").outcome == TaskOutcome.SUCCESS
+    }
+
+
+    def "Check pmd incremental mode enabling on gradle < 6.4"() {
         setup:
         build("""
             plugins {
@@ -23,6 +61,14 @@ class PmdIncrementalAnalysisKitTest extends AbstractKitTest {
             
                 pmdIncremental = true
             }
+            
+             task validator {
+                doFirst() {
+                    if (!pmdMain.incrementalAnalysis.get()) {
+                        throw new IllegalStateException('PMD incremental analysis not enabled');
+                    }
+                }
+            }
 
             repositories {
                 mavenCentral() //required for testKit run
@@ -32,7 +78,7 @@ class PmdIncrementalAnalysisKitTest extends AbstractKitTest {
         fileFromClasspath('src/main/java/sample/Sample.java', '/ru/vyarus/gradle/plugin/quality/java/sample/Sample.java')
 
         when: "run check task with java sources"
-        BuildResult result = run('pmdMain')
+        BuildResult result = runVer('6.3','pmdMain', 'validator')
 
         then: "all plugins detect violations"
         result.task(":pmdMain").outcome == TaskOutcome.SUCCESS