+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+* Deprecate quality.pmdIncremental property because incremental analysis is enabled by default
+    since gradle 6.4 (property only useful for enabling it in gradle 5.6 - 6.3)
+
 ### 4.6.0 (2021-05-01)
 * Fix gradle 7 compatibility (for checkstyle plugin)
 * Fix console output on windows (remove special unicode separator characters, appearing incorrectly)